@@ -28,103 +29,63 @@
     "usePathForSuiteName": "true"
   },
   "dependencies": {
-    "amqp-ts": "1.5.0",
-    "amqplib": "0.5.3",
-    "body-parser": "1.19.0",
-    "class-transformer": "0.2.3",
-    "class-transformer-validator": "0.8.0",
-    "cloud-config-client": "1.4.2",
-    "config": "3.2.2",
-    "connect-redis": "3.4.2",
-    "cookie-parser": "1.4.5",
-    "copy-to": "2.0.1",
-    "cors": "2.8.5",
-    "express": "4.17.1",
-    "express-session": "1.16.2",
-    "express-winston": "3.2.1",
-    "glob": "7.1.4",
-    "helmet": "3.18.0",
-    "js-base64": "2.5.2",
-    "js-yaml": "3.13.1",
-    "log4js": "6.1.2",
-    "moment": "2.26.0",
-    "moment-timezone": "0.5.31",
-    "mongodb": "3.2.7",
-    "mongoose": "5.6.6",
-    "node-fetch": "2.6.0",
-    "node-health-service": "1.0.7",
-    "node-rest-client": "3.1.0",
-    "pg": "7.12.1",
-    "redis": "2.8.0",
     "request": "2.88.2",
+    "body-parser": "1.20.2",
+    "class-transformer": "0.5.1",
+    "class-transformer-validator": "0.9.1",
+    "cloud-config-client": "1.6.2",
+    "config": "3.3.9",
+    "cookie-parser": "1.4.6",
+    "express": "4.18.2",
+    "helmet": "7.0.0",
+    "js-base64": "3.7.5",
+    "log4js": "6.9.1",
+    "moment-timezone": "0.5.43",
+    "node-health-service": "1.0.7",
+    "pg": "^8.11.0",
     "request-context": "2.0.0",
-    "routing-controllers": "0.7.7",
-    "sprintf-js": "1.1.2",
-    "swagger-ui-express": "4.0.7",
-    "typed-rest-client": "1.5.0",
+    "routing-controllers": "0.10.4",
+    "sprintf-js": "1.1.3",
+    "swagger-ui-express": "5.0.0",
     "typedi": "0.8.0",
-    "typeorm": "0.2.24",
-    "typeorm-typedi-extensions": "0.2.3",
-    "uuid": "3.3.2",
-    "winston": "3.2.1",
-    "zipkin": "0.18.6",
-    "zipkin-context-cls": "0.18.5",
-    "zipkin-instrumentation-express": "0.18.6",
-    "zipkin-transport-http": "0.18.6"
+    "typeorm": "^0.3.17",
+    "uuid": "9.0.1",
+    "class-validator": "0.14.0",
+    "js-yaml": "4.1.0"
   },
   "devDependencies": {
-    "@types/jest": "25.2.3",
-    "@types/sprintf-js": "1.1.2",
-    "@types/js-base64": "2.3.1",
-    "@types/amqplib": "0.5.13",
-    "@types/bluebird": "3.5.27",
-    "@types/body-parser": "1.17.0",
-    "@types/config": "0.0.34",
-    "@types/connect-redis": "0.0.10",
-    "@types/cookie-parser": "1.4.2",
-    "@types/cors": "2.8.5",
-    "@types/express": "4.17.0",
-    "@types/express-session": "1.15.13",
-    "@types/glob": "7.1.1",
-    "@types/helmet": "0.0.35",
-    "@types/js-yaml": "3.9.0",
-    "@types/log4js": "2.3.5",
-    "@types/moment": "2.13.0",
-    "@types/moment-timezone": "0.5.13",
-    "@types/mongodb": "3.1.30",
-    "@types/mongoose": "5.5.9",
-    "@types/node": "12.6.8",
-    "@types/pg": "7.11.2",
-    "@types/redis": "2.8.13",
     "@types/request": "2.48.5",
-    "@types/supertest": "2.0.8",
-    "@types/swagger-ui-express": "3.0.1",
-    "@types/uuid": "3.4.5",
-    "@typescript-eslint/eslint-plugin": "2.6.1",
-    "@typescript-eslint/parser": "2.6.1",
-    "eslint": "6.6.0",
-    "eslint-config-prettier": "6.5.0",
-    "eslint-config-standard": "14.1.0",
-    "eslint-loader": "3.0.2",
-    "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-jest": "23.8.1",
-    "eslint-plugin-node": "10.0.0",
-    "eslint-plugin-prettier": "3.1.1",
-    "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-standard": "4.0.1",
-    "istanbul": "0.4.5",
-    "jest": "26.0.1",
-    "jest-html-reporter": "2.8.0",
-    "jest-junit": "10.0.0",
+    "@types/jest": "^29.5.6",
+    "@types/sprintf-js": "1.1.2",
+    "@types/js-base64": "3.0.0",
+    "@types/body-parser": "1.19.4",
+    "@types/config": "3.3.1",
+    "@types/cookie-parser": "1.4.5",
+    "@types/express": "4.17.20",
+    "@types/helmet": "4.0.0",
+    "@types/node": "^18.16.10",
+    "@types/pg": "8.10.5",
+    "@types/supertest": "2.0.14",
+    "@types/swagger-ui-express": "4.1.4",
+    "@types/uuid": "9.0.5",
+    "@typescript-eslint/eslint-plugin": "6.8.0",
+    "@typescript-eslint/parser": "6.8.0",
+    "eslint": "8.51.0",
+    "eslint-config-standard": "17.1.0",
+    "eslint-plugin-import": "2.28.1",
+    "eslint-plugin-jest": "27.4.2",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-standard": "4.1.0",
+    "eslint-plugin-n": "^16.2.0",
+    "jest": "^29.7.0",
+    "jest-html-reporter": "^3.10.2",
+    "jest-junit": "16.0.0",
     "npm-run-all": "4.1.5",
-    "prettier": "1.19.0",
-    "remap-istanbul": "0.13.0",
-    "rimraf": "2.6.3",
-    "supertest": "4.0.2",
-    "ts-jest": "26.1.1",
-    "ts-loader": "6.2.1",
-    "ts-node": "8.3.0",
-    "typescript": "3.9.5",
-    "xunit-file": "1.0.0"
+    "rimraf": "5.0.5",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.1",
+    "ts-node": "10.9.1",
+    "typescript": "^4.9.5"
   }
 }
