@@ -34,11 +34,11 @@ export class CloudConfig {
                 name: this.configName
             }).then((config: any) => {
                 this.configData = config;
-                resolve();
+                resolve(null);
             }).catch((err: any) => {
                 systemLogger.error(getBuildInfo(), err);
                 reject(err);
             });
         });
-    }
+    };
 }
